Modularise functionality of plot_corners.py.
